+++ engine/runtime-suggestions/src/main/scala/org/enso/compiler/suggestions/SuggestionBuilder.scala
@@ -94,6 +94,7 @@ final class SuggestionBuilder[A: IndexedSource](
                     annotations,
                     _,
                     _,
+                    _,
                     _
                   ) =>
                 buildAtomConstructor(